fix security issues

+++ src/main/java/org/infai/ses/senergy/utils/StreamsConfigProvider.java
@@ -42,7 +42,7 @@ public class StreamsConfigProvider {
         streamsConfiguration.put(StreamsConfig.DEFAULT_TIMESTAMP_EXTRACTOR_CLASS_CONFIG, WallclockTimestampExtractor.class);
         streamsConfiguration.put(StreamsConfig.NUM_STREAM_THREADS_CONFIG, Helper.getEnv("STREAM_THREADS_CONFIG", "1"));
         streamsConfiguration.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, Helper.getEnv("CONSUMER_AUTO_OFFSET_RESET_CONFIG", "earliest"));
-        streamsConfiguration.put(StreamsConfig.CACHE_MAX_BYTES_BUFFERING_CONFIG, Integer.parseInt(Helper.getEnv("CACHE_MAX_BYTES_BUFFERING_CONFIG", "10")) * 1024 * 1024L);
+        streamsConfiguration.put(StreamsConfig.CACHE_MAX_BYTES_BUFFERING_CONFIG, Integer.parseInt(Helper.getEnv("CACHE_MAX_BYTES_BUFFERING_CONFIG", "10")) *(long) 1024 * 1024L);
         streamsConfiguration.put(StreamsConfig.COMMIT_INTERVAL_MS_CONFIG, Integer.parseInt(Helper.getEnv("COMMIT_INTERVAL_MS_CONFIG", "1000")));
     }
